pyproject.toml
[tool.poetry]
name = "ml-workstation-server"
version = "0.1.0"
description = ""
authors = ["mr-raccoon-97 <eric.m.cardozo@mi.unc.edu.ar>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
aio-pika = "^9.4.3"
uvicorn = "^0.31.1"
pydantic-settings = "^2.5.2"
motor = "^3.6.0"
aiohttp = "^3.10.10"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
